Moved isDateWithinFiltersDateRange function to dateHelpers.js
